Refactored to allow better handling of different environments/branches.  Allows better use of louisconf.py configuration settings.

+++ louis/commands/packages.py
@@ -86,6 +86,7 @@ def patch_virtualenv(user, package_path, virtualenv_path='env'):
     Symlinks package_path in virtual env's site-packages.
     """
     with settings(user=user):
-        target = '/home/%s/%s/lib/python2.6/site-packages/' % 
-                 (user, virtualenv_path)
-        run('ln -s %s %s' % (package_path, target))
+        target = ('/home/%s/%s/lib/python2.6/site-packages/' % 
+                  (user, virtualenv_path))
+        if not files.exists(target, use_sudo=False):
+            run('ln -s %s %s' % (package_path, target))
